chore: put kuma_log_context in pkg/log/logger.go file
