+++ pkg/core/managers/apis/mesh/mesh_manager.go
@@ -14,7 +14,7 @@ import (
 	core_registry "github.com/kumahq/kuma/pkg/core/resources/registry"
 	core_store "github.com/kumahq/kuma/pkg/core/resources/store"
 	defaults_mesh "github.com/kumahq/kuma/pkg/defaults/mesh"
-	kuma_log_context "github.com/kumahq/kuma/pkg/log/context"
+	kuma_log "github.com/kumahq/kuma/pkg/log"
 )
 
 func NewMeshManager(
@@ -61,7 +61,7 @@ func (m *meshManager) List(ctx context.Context, list core_model.ResourceList, fs
 }
 
 func (m *meshManager) Create(ctx context.Context, resource core_model.Resource, fs ...core_store.CreateOptionsFunc) error {
-	ctx = kuma_log_context.NewContext(ctx, kuma_log_context.FromContext(ctx))
+	ctx = kuma_log.NewContext(ctx, kuma_log.FromContext(ctx))
 	opts := core_store.NewCreateOptions(fs...)
 	mesh, err := m.mesh(resource)
 	if err != nil {
