+++ pkg/defaults/components.go
@@ -18,7 +18,7 @@ import (
 	"github.com/kumahq/kuma/pkg/core/runtime/component"
 	"github.com/kumahq/kuma/pkg/core/tokens"
 	"github.com/kumahq/kuma/pkg/core/user"
-	kuma_log_context "github.com/kumahq/kuma/pkg/log/context"
+	kuma_log "github.com/kumahq/kuma/pkg/log"
 	"github.com/kumahq/kuma/pkg/tokens/builtin/zone"
 	"github.com/kumahq/kuma/pkg/tokens/builtin/zoneingress"
 )
@@ -36,10 +36,10 @@ func Setup(runtime runtime.Runtime) error {
 		zoneIngressSigningKeyManager := tokens.NewSigningKeyManager(runtime.ResourceManager(), zoneingress.ZoneIngressSigningKeyPrefix)
 
 		ctx := runtime.AppContext()
-		logger := kuma_log_context.FromContext(ctx).
+		logger := kuma_log.FromContext(ctx).
 			WithName("defaults").
 			WithValues("secretPrefix", zoneingress.ZoneIngressSigningKeyPrefix)
-		ctx = kuma_log_context.NewContext(ctx, logger)
+		ctx = kuma_log.NewContext(ctx, logger)
 
 		if err := runtime.Add(tokens.NewDefaultSigningKeyComponent(ctx, zoneIngressSigningKeyManager)); err != nil {
 			return err
@@ -89,7 +89,7 @@ func (d *defaultsComponent) NeedLeaderElection() bool {
 
 func (d *defaultsComponent) Start(stop <-chan struct{}) error {
 	// todo(jakubdyszkiewicz) once this https://github.com/kumahq/kuma/issues/1001 is done. Wait for all the components to be ready.
-	ctx := kuma_log_context.NewContext(user.Ctx(context.Background(), user.ControlPlane), core.Log)
+	ctx := kuma_log.NewContext(user.Ctx(context.Background(), user.ControlPlane), core.Log)
 	ctx, cancelFn := context.WithCancel(ctx)
 	defer cancelFn()
 	wg := &sync.WaitGroup{}
