+++ pkg/plugins/runtime/k8s/controllers/mesh_defaults_controller.go
@@ -13,7 +13,7 @@ import (
 	core_model "github.com/kumahq/kuma/pkg/core/resources/model"
 	"github.com/kumahq/kuma/pkg/core/resources/store"
 	defaults_mesh "github.com/kumahq/kuma/pkg/defaults/mesh"
-	kuma_log_context "github.com/kumahq/kuma/pkg/log/context"
+	kuma_log "github.com/kumahq/kuma/pkg/log"
 	common_k8s "github.com/kumahq/kuma/pkg/plugins/common/k8s"
 	"github.com/kumahq/kuma/pkg/plugins/resources/k8s"
 	mesh_k8s "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/api/v1alpha1"
@@ -26,7 +26,7 @@ type MeshDefaultsReconciler struct {
 }
 
 func (r *MeshDefaultsReconciler) Reconcile(ctx context.Context, req kube_ctrl.Request) (kube_ctrl.Result, error) {
-	ctx = kuma_log_context.NewContext(ctx, r.Log)
+	ctx = kuma_log.NewContext(ctx, r.Log)
 	mesh := core_mesh.NewMeshResource()
 	if err := r.ResourceManager.Get(ctx, mesh, store.GetByKey(req.Name, core_model.NoMesh)); err != nil {
 		if kube_apierrs.IsNotFound(err) {
